Fixed a case where the uses version would not be normalized in the case of linting
